Add marker watermark on cell hovering
In order to show a watermark and animate the stroke I duplicated the
elements inside the svg cross' and circles, to be able to add
independent transitions to each element. The watermark shows when the
cell is hovered, then it disappears instantaneously on a click
and finally the strokes are animated.

I tried to achieve the same result using a single element, but I found
difficult to tweak the transitions.