Add stroke animation to AI move
Previously the markers hade the display property set to none.
The display changes to inline on mouse over, and then on click the
animation is triggered. When the AI plays, the display remains set to
none, because there is no mouse over event this prevents the animation
to be shown. I tried using setTimeOut to add a delay and dispatch a
mouseover event or directly set the display of the marker to inline.
The issue is that the flow of the code is not linear, which is probably
preventing the animation to be rendered. Currently I changed the display
property of the markers to inline and I use a dash-offset to hide the
markers instead, the issue is that now the markers occupy a different
position.

I can solve this issue setting the position of the markers to relative
to the div container. Another option would be to use a variable that
prevents the marker to get the class 'clicked' which triggers the
animation, before it getting the class 'over' which displays the marker.