Fix markers position
I decided to make the position of the markers absolute relative to the
cell container. I also tried the approached mention in the previous
commit where I add the class 'clicked' only after the cell has the class
'over', but the animation still didn't play.

There was another bug, in which the cells played by the AI where not
being added to the class 'clicked' causing subsequent moves to pick
already picked cell.