Refactor game logic
To be able to play different types of games I refactor the original
code to use a game object and a player object generator that creates
user type and computer type players. I integrated the functions for
getting the user and AI moves into a single function getPlayerMove
that it;s encapsulated within the game object.

Previously the getPlayerMove called the getAIMove and the flow of the
game continued jumping back and forth between AI and user turns. In
order to enable a two player mode, I am trying to use promises to
wait indefinitely until one of the players make a move.

Currently it seems that the logic of the game works but how I used to
show the markers is causing some problems.