Fix rendering bug
The reason why the markers where not rendered is because I had some
query selectors that used the variable markers, which contained the
strings "circle" and "cross". The query selector switch between strings
based on wether the current turn was odd or even. I change this logic
to use the marker stored in the player object.

Now one game is correctly rendered, but the game is blocked after
a game finishes. This behavior is likely to be the result of a promise
being unresolved.