Fix restart button bug
The restart button changed a flag that was used to break the loop. The
problem was that the flow inside the loop is being blocked by a Promise
which is only resolved when a game finishes. I change the event listener
function attached to the restart button so that upon click it forces
the page to reload, which achieves the desired behavior of resetting the
game.